LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/next-hop-protocol.https-expected.txt


FAIL nextHopProtocol reports H1 correctly when routed via a service worker. assert_equals: nextHopProtocol is set on pass-through expected "http/1.1" but got ""
FAIL nextHopProtocol reports H2 correctly when routed via a service worker. assert_equals: nextHopProtocol is set on fallback expected "h2" but got "http/1.1"
